+++ .github/workflows/build.yml
@@ -299,6 +299,7 @@ jobs:
               "PyBuilder/jacoco" \
               "WorkflowCore/jacoco" \
               "WorkflowOperator/jacoco" \
+              "WorkflowCompiler/jacoco" \
               "WorkflowExecutionService/jacoco"
       - name: Upload amber and common coverage to Codecov
         if: always()
