+++ common/workflow-compiler/src/main/scala/org/apache/texera/common/compiler/model/LogicalLink.scala
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.texera.amber.compiler.model
+package org.apache.texera.common.compiler.model
 
 import com.fasterxml.jackson.annotation.{JsonCreator, JsonProperty}
 import org.apache.texera.amber.core.virtualidentity.OperatorIdentity
@@ -29,6 +29,19 @@ case class LogicalLink(
     @JsonProperty("toOpId") toOpId: OperatorIdentity,
     toPortId: PortIdentity
 ) {
+  require(
+    fromOpId != null && fromOpId.id != null && fromOpId.id.nonEmpty,
+    "LogicalLink fromOpId must be non-null and non-empty"
+  )
+  require(
+    toOpId != null && toOpId.id != null && toOpId.id.nonEmpty,
+    "LogicalLink toOpId must be non-null and non-empty"
+  )
+  require(
+    fromOpId != toOpId,
+    s"LogicalLink self-loop not allowed: fromOpId == toOpId == ${fromOpId.id}"
+  )
+
   @JsonCreator
   def this(
       @JsonProperty("fromOpId") fromOpId: String,
